% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_to_robj.R
\name{response_to_robj}
\alias{response_to_robj}
\title{response_to_robj}
\usage{
response_to_robj(url, drop_data = FALSE, query = list(), fun = GET)
}
\arguments{
\item{url}{The location of the api endpoint, excluding host and port}
}
\description{
response_to_robj
}
