% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-package.R
\docType{package}
\name{state-package}
\alias{state-package}
\title{state: State Management for Reactive Contexts}
\description{
Utilities to define typed reactive values for shiny applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calderonsamuel/state}
  \item \url{https://calderonsamuel.github.io/state/}
  \item Report bugs at \url{https://github.com/calderonsamuel/state/issues}
}

}
\author{
\strong{Maintainer}: Samuel Calderon \email{samuel.calderon@uarm.pe}

}
\keyword{internal}
