% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_manager.R
\name{state_manager}
\alias{state_manager}
\title{Typed reactiveValues}
\usage{
state_manager(...)
}
\arguments{
\item{...}{Objects that will be added to the reactivevalues object. All of
these objects must be named.}
}
\description{
A \code{shiny::reactiveValues} object with type safety features.
All elements must be named and initialized with \code{state()}
}
