% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state-package.R
\docType{package}
\name{state-package}
\alias{state}
\alias{state-package}
\title{state: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/calderonsamuel/state}
  \item \url{https://calderonsamuel.github.io/state/}
  \item Report bugs at \url{https://github.com/calderonsamuel/state/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
