% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{state}
\alias{state}
\title{State}
\usage{
state(value = NULL, type = NULL)
}
\arguments{
\item{value}{Initial value of the state}

\item{type}{The expected class}
}
\description{
A way of defining state for a shiny application. Alternative to \code{shiny::reactiveValues()}.
}
